+++ CHANGELOG.md
@@ -8,7 +8,8 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 
 ### Added
 
-- Added a machine translated Tagalog locale (#1089).
+- We added a machine translated Tagalog locale (#1089).
+- We added additional keywords to all emojis to improve searching (#1090), but also translated all of those keywords, so that works in all supported locales (#639).
 
 ### Changed